Before
===
```
% ./test/benchmark
nanoid                      535,348 ops/sec
customAlphabet              515,498 ops/sec
uid.sync                    292,991 ops/sec
uuid v4                   1,383,186 ops/sec
secure-random-string        288,087 ops/sec
cuid                        144,727 ops/sec
shortid                      48,198 ops/sec

Async:
async nanoid                 70,750 ops/sec
async customAlphabet         69,236 ops/sec
async secure-random-string   63,352 ops/sec
uid                          61,122 ops/sec

Non-secure:
non-secure nanoid         2,813,255 ops/sec
rndm                      2,663,304 ops/sec
```

After
===
```
% ./test/benchmark
nanoid                    2,094,623 ops/sec             <-- ~4x faster
customAlphabet            1,764,130 ops/sec             <-- ~3.5x faster
uid.sync                    294,920 ops/sec
uuid v4                   1,389,111 ops/sec
secure-random-string        286,511 ops/sec
cuid                        136,146 ops/sec
shortid                      47,159 ops/sec

Async:
async nanoid                 68,758 ops/sec
async customAlphabet         69,112 ops/sec
async secure-random-string   62,578 ops/sec
uid                          61,228 ops/sec

Non-secure:
non-secure nanoid         2,744,634 ops/sec
rndm                      2,567,437 ops/sec
```

Notes
===

I did not make an effort to apply this technique to the entire codebase, because I didn't know if the maintainer(s) would be amenable to the change. If they are, I can expand this PR to address those parts